requirements.txt
mysqlclient==2.0.3
python-dotenv==0.17.1
requests==2.25.1
